war against 'if': use 'case' instead
This converts all 'if' blocks which contain exactly one
'if COND' clause and one 'else' clause to 'case' blocks.

The possibly two branches which are "true" in some sense
in an 'if' block like

    X = if N > 5 -> foo;
           true  -> bar
        end,

are a little confusing to the reader. In most cases,
Erlang code should use 'case' instead of 'if' for
much more obvious code:

    X = case N > 5 of
            true -> foo;
            false -> bar
        end,

This leaves two 'if' blocks alone in backend_file.erl
due to 'case' actually not fitting the logic there, and
two 'if' blocks in wire_connection.erl which can be handled
better than by using true and false.